Reorganization of the Components into folder for each view generally.
