+++ src/components/Information/CitationsRender.js
