+++ src/components/Information/LicenseRender.js
@@ -23,13 +23,13 @@ const LicenseRender = ({ information }) => {
         let userEndorsements = [];
         for(var index in endorsementArr) {
             console.log(ENDORSEMENT_DICT[endorsementArr[index]]);
-            userEndorsements.push(<li><strong>{endorsementArr[index]}: </strong>{ENDORSEMENT_DICT[endorsementArr[index]]}</li>)
+            userEndorsements.push(<li key={endorsementArr[index]+ "end"}><strong>{endorsementArr[index]}: </strong>{ENDORSEMENT_DICT[endorsementArr[index]]}</li>)
         }
         userEndorsements.push(<li><strong>Other Endorsements</strong></li>);
         for(var key in ENDORSEMENT_DICT) {
             if(endorsementArr.includes(key)) continue;
             else {
-                userEndorsements.push(<li><strong>{key}: </strong>{ENDORSEMENT_DICT[key]}</li>)
+                userEndorsements.push(<li key={key + "end"}><strong>{key}: </strong>{ENDORSEMENT_DICT[key]}</li>)
             }
         }
         let endorsementGlossary = (
@@ -60,13 +60,13 @@ const LicenseRender = ({ information }) => {
         }
         for(var index in restrictionArr) {
             console.log(ENDORSEMENT_DICT[restrictionArr[index]]);
-            userRestrictions.push(<li><strong>{restrictionArr[index]}: </strong>{RESTRICTION_DICT[restrictionArr[index]]}</li>)
+            userRestrictions.push(<li key={restrictionArr[index] + "res"}><strong>{restrictionArr[index]}: </strong>{RESTRICTION_DICT[restrictionArr[index]]}</li>)
         }
         userRestrictions.push(<li><strong>Other Endorsements</strong></li>);
         for(var key in ENDORSEMENT_DICT) {
             if(restrictionArr.includes(key)) continue;
             else {
-                userRestrictions.push(<li><strong>{key}: </strong>{ENDORSEMENT_DICT[key]}</li>)
+                userRestrictions.push(<li key={key + "res"}><strong>{key}: </strong>{ENDORSEMENT_DICT[key]}</li>)
             }
         }
         let endorsementGlossary = (
@@ -120,7 +120,6 @@ const LicenseRender = ({ information }) => {
                 <div className="col">
                     <h3>License Endorsement Glossary</h3>
                     { generateEndorsementGlossary(endorsements)}
-                    Separator
                     { /*generateFullEndorsementGlossary() */}
                 </div>
             </section>
@@ -128,7 +127,6 @@ const LicenseRender = ({ information }) => {
                 <div className="col">
                     <h3>License Restriction Glossary</h3>
                     { generateRestrictionGLossary(restrictions) }
-                    Separator
                     { /*generateFullRestrictionGlossary() */}
                 </div>
             </section>
