+++ src/components/User/Login.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Redirect } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { loginAttempt } from '../actions/userActions';
+import { loginAttempt } from '../../actions/userActions';
 import ErrorBanner from './ErrorBanner';
