+++ src/components/User/Profile.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { updateContactAttempt } from '../actions/profileActions';
+import { updateContactAttempt } from '../../actions/profileActions';
 import SuccessBanner from './SuccessBanner';
 
 const ProfilePage = () => {
